remove many deprecated structures
