+++ rust/tw_bitcoin/src/ffi/scripts.rs
@@ -1,20 +1,16 @@
-use crate::brc20::{BRC20TransferInscription, Ticker};
+use crate::brc20::Ticker;
 use crate::nft::OrdinalNftInscription;
-use crate::{
-    Recipient, TXOutputP2TRScriptPath, TxOutputP2PKH, TxOutputP2TRKeyPath, TxOutputP2WPKH,
-};
+use crate::Recipient;
 use bitcoin::{PublicKey, WPubkeyHash};
-use std::borrow::Cow;
 use std::ffi::{c_char, CStr};
 use tw_memory::ffi::c_byte_array::CByteArray;
 use tw_memory::ffi::c_byte_array_ref::CByteArrayRef;
 use tw_misc::try_or_else;
-use tw_proto::Bitcoin::Proto::TransactionOutput;
 
 #[no_mangle]
 // Builds the P2PKH scriptPubkey.
 pub unsafe extern "C" fn tw_build_p2pkh_script(
-    satoshis: i64,
+    _satoshis: i64,
     pubkey: *const u8,
     pubkey_len: usize,
 ) -> CByteArray {
@@ -23,8 +19,9 @@ pub unsafe extern "C" fn tw_build_p2pkh_script(
         CByteArrayRef::new(pubkey, pubkey_len).as_slice(),
         CByteArray::null
     );
-    let recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
+    let _recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
 
+    /*
     let tx_out = TxOutputP2PKH::new(satoshis as u64, recipient);
 
     // Prepare and serialize protobuf structure.
@@ -37,12 +34,15 @@ pub unsafe extern "C" fn tw_build_p2pkh_script(
     let serialized = tw_proto::serialize(&proto).expect("failed to serialized transaction output");
 
     CByteArray::from(serialized)
+    */
+
+    todo!()
 }
 
 #[no_mangle]
 // Builds the P2WPKH scriptPubkey.
 pub unsafe extern "C" fn tw_build_p2wpkh_script(
-    satoshis: i64,
+    _satoshis: i64,
     pubkey: *const u8,
     pubkey_len: usize,
 ) -> CByteArray {
@@ -51,29 +51,18 @@ pub unsafe extern "C" fn tw_build_p2wpkh_script(
         CByteArrayRef::new(pubkey, pubkey_len).as_slice(),
         CByteArray::null
     );
-    let recipient = try_or_else!(
+    let _recipient = try_or_else!(
         Recipient::<WPubkeyHash>::from_slice(slice),
         CByteArray::null
     );
 
-    let tx_out = TxOutputP2WPKH::new(satoshis as u64, recipient);
-
-    // Prepare and serialize protobuf structure.
-    let proto = TransactionOutput {
-        value: satoshis,
-        script: Cow::from(tx_out.script_pubkey.as_bytes()),
-        spendingScript: Cow::default(),
-    };
-
-    let serialized = tw_proto::serialize(&proto).expect("failed to serialized transaction output");
-
-    CByteArray::from(serialized)
+    todo!()
 }
 
 #[no_mangle]
 // Builds the P2TR key-path scriptPubkey.
 pub unsafe extern "C" fn tw_build_p2tr_key_path_script(
-    satoshis: i64,
+    _satoshis: i64,
     pubkey: *const u8,
     pubkey_len: usize,
 ) -> CByteArray {
@@ -82,20 +71,9 @@ pub unsafe extern "C" fn tw_build_p2tr_key_path_script(
         CByteArrayRef::new(pubkey, pubkey_len).as_slice(),
         CByteArray::null
     );
-    let recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
+    let _recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
 
-    let tx_out = TxOutputP2TRKeyPath::new(satoshis as u64, recipient.into());
-
-    // Prepare and serialize protobuf structure.
-    let proto = TransactionOutput {
-        value: satoshis,
-        script: Cow::from(tx_out.script_pubkey.as_bytes()),
-        spendingScript: Cow::default(),
-    };
-
-    let serialized = tw_proto::serialize(&proto).expect("failed to serialized transaction output");
-
-    CByteArray::from(serialized)
+    todo!()
 }
 
 #[no_mangle]
@@ -103,8 +81,8 @@ pub unsafe extern "C" fn tw_build_p2tr_key_path_script(
 pub unsafe extern "C" fn tw_build_brc20_transfer_inscription(
     // The 4-byte ticker.
     ticker: *const c_char,
-    amount: u64,
-    satoshis: i64,
+    _amount: u64,
+    _satoshis: i64,
     pubkey: *const u8,
     pubkey_len: usize,
 ) -> CByteArray {
@@ -118,7 +96,7 @@ pub unsafe extern "C" fn tw_build_brc20_transfer_inscription(
         return CByteArray::null();
     }
 
-    let ticker = Ticker::new(ticker.to_string()).expect("ticker must be 4 bytes");
+    let _ticker = Ticker::new(ticker.to_string()).expect("ticker must be 4 bytes");
 
     // Convert Recipient
     let slice = try_or_else!(
@@ -126,25 +104,20 @@ pub unsafe extern "C" fn tw_build_brc20_transfer_inscription(
         CByteArray::null
     );
 
-    let recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
-
-    // Build transfer inscription.
-    let transfer = BRC20TransferInscription::new(recipient, ticker, amount)
-        .expect("transfer inscription implemented wrongly");
-
-    let tx_out = TXOutputP2TRScriptPath::new(satoshis as u64, transfer.inscription().recipient());
-    let spending_script = transfer.inscription().taproot_program();
+    let _recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
 
     // Prepare and serialize protobuf structure.
+    /*
+    let spending_script = transfer.inscription().taproot_program();
+
     let proto = TransactionOutput {
         value: satoshis,
         script: Cow::from(tx_out.script_pubkey.as_bytes()),
         spendingScript: Cow::from(spending_script.as_bytes()),
     };
+    */
 
-    let serialized = tw_proto::serialize(&proto).expect("failed to serialized transaction output");
-
-    CByteArray::from(serialized)
+    todo!()
 }
 
 #[no_mangle]
@@ -153,7 +126,7 @@ pub unsafe extern "C" fn tw_bitcoin_build_nft_inscription(
     mime_type: *const c_char,
     data: *const u8,
     data_len: usize,
-    satoshis: i64,
+    _satoshis: i64,
     pubkey: *const u8,
     pubkey_len: usize,
 ) -> CByteArray {
@@ -178,20 +151,8 @@ pub unsafe extern "C" fn tw_bitcoin_build_nft_inscription(
     let recipient = try_or_else!(Recipient::<PublicKey>::from_slice(slice), CByteArray::null);
 
     // Inscribe NFT data.
-    let nft = OrdinalNftInscription::new(mime_type.as_bytes(), data, recipient)
+    let _nft = OrdinalNftInscription::new(mime_type.as_bytes(), data, recipient)
         .expect("Ordinal NFT inscription incorrectly constructed");
 
-    let tx_out = TXOutputP2TRScriptPath::new(satoshis as u64, nft.inscription().recipient());
-    let spending_script = nft.inscription().taproot_program();
-
-    // Prepare and serialize protobuf structure.
-    let proto = TransactionOutput {
-        value: satoshis,
-        script: Cow::from(tx_out.script_pubkey.as_bytes()),
-        spendingScript: Cow::from(spending_script.as_bytes()),
-    };
-
-    let serialized = tw_proto::serialize(&proto).expect("failed to serialized transaction output");
-
-    CByteArray::from(serialized)
+    todo!()
 }
