--- rust/tw_bitcoin/src/tests/ffi/mod.rs
@@ -1,5 +1,0 @@
-mod brc20_transfer;
-mod nft;
-mod scripts;
-mod transaction;
-mod utils;
