+++ rust/tw_bitcoin/src/tests/mod.rs
@@ -1,11 +1,3 @@
 mod address;
-mod brc20_transfer;
-mod data;
-mod ffi;
-mod nft;
-mod p2pkh;
-mod p2tr_key_path;
-mod p2wpkh;
-mod prehashes;
-
-pub const ONE_BTC: u64 = 100_000_000;
+//mod data;
+mod coin_entry_sign;
